refactor: Move lib/* files
